+++ lib/ui/home/HomeAthleticsEventsWidget.dart
@@ -3,13 +3,13 @@ import 'dart:async';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:illinois/model/News.dart';
+import 'package:illinois/model/sport/Game.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Config.dart';
 import 'package:illinois/service/Sports.dart';
-import 'package:illinois/ui/athletics/AthleticsNewsArticlePanel.dart';
-import 'package:illinois/ui/athletics/AthleticsNewsCard.dart';
-import 'package:illinois/ui/athletics/AthleticsNewsListPanel.dart';
+import 'package:illinois/ui/athletics/AthleticsGameDetailPanel.dart';
+import 'package:illinois/ui/athletics/AthleticsHomePanel.dart';
+import 'package:illinois/ui/explore/ExplorePanel.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
 import 'package:illinois/ui/home/HomeWidgets.dart';
 import 'package:illinois/ui/widgets/LinkButton.dart';
@@ -22,25 +22,25 @@ import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/section_header.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
-class HomeAthliticsNewsWidget extends StatefulWidget {
+class HomeAthliticsEventsWidget extends StatefulWidget {
 
   final String? favoriteId;
   final StreamController<String>? updateController;
 
-  HomeAthliticsNewsWidget({Key? key, this.favoriteId, this.updateController}) : super(key: key);
+  HomeAthliticsEventsWidget({Key? key, this.favoriteId, this.updateController}) : super(key: key);
 
   static Widget handle({String? favoriteId, HomeDragAndDropHost? dragAndDropHost, int? position}) =>
     HomeHandleWidget(favoriteId: favoriteId, dragAndDropHost: dragAndDropHost, position: position,
-      title: Localization().getStringEx('widget.home.athletics_news.text.title', 'Athletics News'),
+      title: Localization().getStringEx('widget.home.athletics_events.text.title', 'Athletics Events'),
     );
 
-  State<HomeAthliticsNewsWidget> createState() => _HomeAthleticsNewsWidgetState();
+  State<HomeAthliticsEventsWidget> createState() => _HomeAthleticsEventsWidgetState();
 }
 
-class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> implements NotificationsListener {
+class _HomeAthleticsEventsWidgetState extends State<HomeAthliticsEventsWidget> implements NotificationsListener {
 
-  List<News>? _news;
-  bool _loadingNews = false;
+  List<Game>? _games;
+  bool _loadingGames = false;
   DateTime? _pausedDateTime;
 
   @override
@@ -55,17 +55,17 @@ class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> imple
     if (widget.updateController != null) {
       widget.updateController!.stream.listen((String command) {
         if (command == HomePanel.notifyRefresh) {
-          _refreshNews(showProgress: true);
+          _refreshGames(showProgress: true);
         }
       });
     }
 
     if (Connectivity().isOnline) {
-      _loadingNews = true;
-      Sports().loadNews(null, Config().homeAthleticsNewsCount).then((List<News>? news) {
+      _loadingGames = true;
+      Sports().loadGames(limit: Config().homeAthleticsEventsCount).then((List<Game>? games) {
         setStateIfMounted(() {
-          _news = news;
-          _loadingNews = false;
+          _games = games;
+          _loadingGames = false;
         });
       });
     }
@@ -84,7 +84,7 @@ class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> imple
   @override
   void onNotification(String name, dynamic param) {
     if (name == Connectivity.notifyStatusChanged) {
-      _refreshNews();
+      _refreshGames();
     }
     else if (name == AppLivecycle.notifyStateChanged) {
       _onAppLivecycleStateChanged(param);
@@ -99,11 +99,11 @@ class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> imple
   @override
   Widget build(BuildContext context) {
     return HomeSlantWidget(favoriteId: widget.favoriteId,
-        title: Localization().getStringEx('widget.home.athletics_news.text.title', 'Athletics News'),
-        titleIcon: Image.asset('images/icon-news.png'),
-        flatHeight: 0, slantHeight: 0,
-        childPadding: EdgeInsets.zero,
-        child: _buildContent(),
+      title: Localization().getStringEx('widget.home.athletics_events.text.title', 'Athletics Events'),
+      titleIcon: Image.asset('images/icon-calendar.png'),
+      flatHeight: 0, slantHeight: 0,
+      childPadding: EdgeInsets.zero,
+      child: _buildContent(),
     );
   }
 
@@ -111,14 +111,14 @@ class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> imple
     if (Connectivity().isOffline) {
       return _buildOfflineContent();
     }
-    else if (_loadingNews) {
+    else if (_loadingGames) {
       return _buildLoadingContent();
     }
-    else if (CollectionUtils.isEmpty(_news)) {
+    else if (CollectionUtils.isEmpty(_games)) {
       return _buildEmptyContent();
     }
     else {
-      return _buildNewsContent();
+      return _buildEventsContent();
     }
 
   }
@@ -128,16 +128,16 @@ class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> imple
       Column(children: <Widget>[
         Text(Localization().getStringEx("app.offline.message.title", "You appear to be offline"), style: TextStyle(fontFamily: Styles().fontFamilies?.bold, fontSize: 20, color: Styles().colors?.fillColorPrimary),),
         Container(height:8),
-        Text(Localization().getStringEx("widget.home.athletics_news.text.offline", "Athletics News are not available while offline"), style: TextStyle(fontFamily: Styles().fontFamilies?.regular, fontSize: 16, color: Styles().colors?.textBackground),),
+        Text(Localization().getStringEx("widget.home.athletics_events.text.offline", "Athletics Events are not available while offline"), style: TextStyle(fontFamily: Styles().fontFamilies?.regular, fontSize: 16, color: Styles().colors?.textBackground),),
     ],),);
   }
 
   Widget _buildEmptyContent() {
     return Padding(padding: EdgeInsets.only(left: 32, right: 32, top: 48, bottom: 48), child:
       Column(children: [
-        Text(Localization().getStringEx("widget.home.athletics_news.text.empty", "Whoops! Nothing to see here."), style: TextStyle(fontFamily: Styles().fontFamilies?.bold, fontSize: 20, color: Styles().colors?.fillColorPrimary),),
+        Text(Localization().getStringEx("widget.home.athletics_events.text.empty", "Whoops! Nothing to see here."), style: TextStyle(fontFamily: Styles().fontFamilies?.bold, fontSize: 20, color: Styles().colors?.fillColorPrimary),),
         Container(height:8),
-        Text(Localization().getStringEx("widget.home.athletics_news.text.empty.description", "No Athletics News are available right now."), style: TextStyle(fontFamily: Styles().fontFamilies?.regular, fontSize: 16, color: Styles().colors?.textBackground),),
+        Text(Localization().getStringEx("widget.home.athletics_events.text.empty.description", "No Athletics Events are available right now."), style: TextStyle(fontFamily: Styles().fontFamilies?.regular, fontSize: 16, color: Styles().colors?.textBackground),),
       ],)
     );
   }
@@ -152,20 +152,21 @@ class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> imple
     );
   }
 
-  Widget _buildNewsContent() {
+  Widget _buildEventsContent() {
     List<Widget> contentList = [];
-    if (_news != null) {
-      for (News news in _news!) {
-        if (contentList.isEmpty && StringUtils.isNotEmpty(news.imageUrl)) {
+    if (_games != null) {
+      for (Game game in _games!) {
+        //contentList.add(AthleticsCard(game: game, onTap: () => _onTapGame(game), showImage: contentList.isEmpty, showInterests: false,),);
+        if (contentList.isEmpty && StringUtils.isNotEmpty(game.imageUrl)) {
           contentList.add(ImageSlantHeader(
-            slantImageColor: Styles().colors!.fillColorPrimaryTransparent03,
-            slantImageAsset: 'images/slant-down-right-blue.png',
-            child: _buildNewsCard(news),
-            imageUrl: news.imageUrl
+            slantImageColor: Styles().colors!.fillColorSecondaryTransparent05,
+            slantImageAsset: 'images/slant-down-right.png',
+            child: AthleticsCard(game: game, onTap: () => _onTapGame(game), showInterests: true,),
+            imageUrl: game.imageUrl
           ));
         }
         else {
-          contentList.add(_buildNewsCard(news),);
+          contentList.add(AthleticsCard(game: game, onTap: () => _onTapGame(game), showInterests: true),);
         }
       }
       contentList.add(
@@ -180,37 +181,37 @@ class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> imple
     return Column(children: contentList,);
   }
 
-  Widget _buildNewsCard(News news) {
-    return Padding(padding: EdgeInsets.only(top: 16, left: 16, right: 16), child:
-      AthleticsNewsCard(news: news, onTap: () => _onTapNews(news)),
-    );
-  }
 
-  void _onTapNews(News news) {
-    Analytics().logSelect(target: "news: "+news.title!);
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsNewsArticlePanel(article: news)));
+  void _onTapGame(Game game) {
+    Analytics().logSelect(target: "Game: "+game.title);
+    if (Connectivity().isNotOffline) {
+      Navigator.push(context, CupertinoPageRoute( builder: (context) => AthleticsGameDetailPanel(game: game)));
+    }
+    else {
+      AppAlert.showOfflineMessage(context, Localization().getStringEx('panel.browse.label.offline.game', 'Game detail is not available while offline.'));
+    }
   }
 
   void _onTapSeeAll() {
-    Analytics().logSelect(target: "HomeAthleticsNews: View All");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsNewsListPanel()));
+    Analytics().logSelect(target: "HomeAthleticsEvents: View All");
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => ExplorePanel(initialItem: ExploreItem.Events, initialFilter: ExploreFilter(type: ExploreFilterType.categories, selectedIndexes: {3}))));
   }
 
-  void _refreshNews({bool showProgress = false}) {
+  void _refreshGames({bool showProgress = false}) {
     if (Connectivity().isOnline) {
       if (showProgress && mounted) {
         setState(() {
-          _loadingNews = true;
+          _loadingGames = true;
         });
       }
-      Sports().loadNews(null, Config().homeAthleticsNewsCount).then((List<News>? news) {
+      Sports().loadGames(limit: Config().homeAthleticsEventsCount).then((List<Game>? games) {
         if (mounted) {
           setState(() {
             if (showProgress) {
-              _loadingNews = false;
+              _loadingGames = false;
             }
-            if (news != null) {
-              _news = news;
+            if (games != null) {
+              _games = games;
             }
           });
         }
@@ -226,7 +227,7 @@ class _HomeAthleticsNewsWidgetState extends State<HomeAthliticsNewsWidget> imple
       if (_pausedDateTime != null) {
         Duration pausedDuration = DateTime.now().difference(_pausedDateTime!);
         if (Config().refreshTimeout < pausedDuration.inSeconds) {
-          _refreshNews();
+          _refreshGames();
         }
       }
     }